Create asan.yml (#131)
Add a GitHub action for ASAN

+++ tests/read_cb.c
@@ -82,7 +82,7 @@ int main(int argc, char *argv[])
 
   if (s3port)
   {
-    int port = atol(s3port);
+    int port = atoi(s3port);
     ms3_set_option(ms3, MS3_OPT_PORT_NUMBER, &port);
   }
